+++ docs/STEP2.md
@@ -7,9 +7,10 @@
     - [X] 사용자
       - [X] [예외] '모든 결과'를 의미하는 'all'이란 이름을 가지는 참가자는 존재할 수 없다.
   - 사다리 게임
-    - [ ] 사다리 결과
+    - [ ] 사다리 결과(End)
+      - [ ] [예외] 결과의 길이는 1글자 이상 5글자 이하여야 한다.
       - [ ] [예외] '꽝' 또는 만 이하의 자연수가 아닌 결과는 존재할 수 없다.
-    - [ ] 사다리 결과들: 사다리 결과의 일급 컬렉션
+    - [ ] 사다리 결과들: 사다리 결과의 일급 컬렉션(Ends)
     - [ ] 인접 경로(enum), 특정 위치의 가로 혹은 세로의 이동 가능 경로 표현
     - [ ] 위치
       - [ ] 현재 높이의 인접 경로에 따라 가로 이동 후 세로 높이를 증가시켜야 한다.
