+++ src/main/java/domain/ladder/LadderWidth.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 public class LadderWidth {
 	private final static String LADDER_WIDTH_NUMBER_ERROR_MSG = "너비는 1 이상이여야 합니다.";
