+++ src/main/java/domain/user/User.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.user;
 
 import java.util.List;
