+++ src/test/java/domain/ladder/LadderBuilderTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -13,7 +13,7 @@ import domain.util.PointGenerator;
 
 class LadderBuilderTest {
 
-	class PresentPointGenerator implements PointGenerator {
+	static class PresentPointGenerator implements PointGenerator {
 		@Override
 		public Point generate() {
 			return Point.PRESENCE;
