+++ src/test/java/domain/user/UserTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.user;
 
 import static org.assertj.core.api.Assertions.*;
